+++ src/storage/store/prefix.rs
@@ -1,4 +1,4 @@
-use derive_more::{Display, From};
+use derive_more::Display;
 use std::path::Path;
 use thiserror::Error;
 
@@ -7,7 +7,7 @@ use crate::node::NodePath;
 /// A Zarr abstract store prefix.
 ///
 /// See <https://zarr-specs.readthedocs.io/en/latest/v3/core/v3.0.html#abstract-store-interface>.
-#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Display, From)]
+#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Display)]
 pub struct StorePrefix(String);
 
 /// An invalid store prefix.